Fixed refactor for get_make_up_options
